blog/src/requirements/base.txt
Django==3.2.5
django-ckeditor==6.1.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1